:hammer: chore: wrap in custom function
